+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [unreleased]
 
+## 33.0.7
+
 ### Fixed
 
 Constraint Enumerator could be not correctly iterated multiple times. Now really fixed.